refactor: split DecoderForm into smaller, focused components
- Extract NetworkSelector component (168 lines) for network selection logic
- Extract TransactionHashInput component (93 lines) for transaction fetching
- Extract CalldataInput component (48 lines) for calldata input handling
- Extract DecodeModeSelector component (80 lines) for signature/ABI mode selection
- Create useDecoderForm hook (47 lines) for coordinated decoding business logic
- Reduce main DecoderForm from 413 to 76 lines (82% reduction)
- Follow Single Responsibility Principle with components under 150-line guideline
- Improve maintainability, testability, and reusability while preserving functionality
